Update pulsar.js
I made several changes to how the dashboard displayed data.  I wanted to be able to see 1) what direction the from_route was traveling in as either "inbound" or "outbound."  2) I wanted to see what direction the to_route was traveling in.  3) I wanted to see the name of the stop that the transfer was occurring at.